v0-no auth

+++ src/actions/action.ts
@@ -1,13 +1,21 @@
 "use server";
 import prisma from "@/lib/db";
 import { PetEssentials } from "@/lib/types";
+import { petFormSchema, petIdSchema } from "@/lib/validations";
 import { Pet } from "@prisma/client";
 import { revalidatePath } from "next/cache";
 
 export async function addPet(pet: PetEssentials) {
+  const validatedPet = petFormSchema.safeParse(pet);
+  if (!validatedPet.success) {
+    return {
+      message: "Invalid Pet Data",
+    };
+  }
+
   try {
     await prisma.pet.create({
-      data: pet,
+      data: validatedPet.data,
     });
   } catch (error) {
     return {
@@ -17,13 +25,21 @@ export async function addPet(pet: PetEssentials) {
   revalidatePath("/app", "layout");
 }
 
-export async function editPet(petId: Pet["id"], newPetData: PetEssentials) {
+export async function editPet(petId: unknown, newPetData: unknown) {
+  const validatedPetId = petIdSchema.safeParse(petId);
+  const validatedPet = petFormSchema.safeParse(newPetData);
+
+  if (!validatedPetId!.success || !validatedPet.success) {
+    return {
+      message: "Invalid Pet Id",
+    };
+  }
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
-      data: newPetData,
+      data: validatedPet.data,
     });
   } catch (error) {
     return {
@@ -33,11 +49,17 @@ export async function editPet(petId: Pet["id"], newPetData: PetEssentials) {
   revalidatePath("/app", "layout");
 }
 
-export async function checkoutPet(petId: Pet["id"]) {
+export async function checkoutPet(petId: unknown) {
+  const validatedPetId = petIdSchema.safeParse(petId);
+  if (!validatedPetId!.success) {
+    return {
+      message: "Invalid Pet Id",
+    };
+  }
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
   } catch (error) {
